ENGR00142092 MX51/MX53: Upgrade SCC2 test driver to 2.6.38
Starting with 2.6.36 kernel, ioctl file operation is removed.
It must be replced with unlocked_ioctl, which has a different
prototype and require a local locking mechanism to prevent
more than 1 user to access ioctls at the same time.

The ioctl function prototype change is actually under the
linux_port.h header file in the kernel sahara driver folder.
The changes in this patch add the locking mechanism.

Signed-off-by: Anish Trivedi <anish@freescale.com>